retrieve product ID from URL to populate product page data
